Bring every state from selectors As Mentioned in PR Review
