+++ src/selectors/contactsSelector.js
@@ -1,0 +1,4 @@
+export const selectContacts = (state) => state.contacts;
+export const selectContactForm = (state) => state.contactForm;
+export const selectLoading = (state) => state.loading;
+export const selectError = (state) => state.error;
